#209 move layer builder utils into utils/layerBuilders/
Addresses PR review feedback: moves buildTextOverlayLayer out of
GeoSetLayer/layerBuilders/ into utils/layerBuilders/ alongside
related polygon utils (buildPolygonLayers, expandPolygonFeatures,
getPointsFromPolygon). Updates all imports accordingly.
