+++ superset-frontend/plugins/geoset-map-chart/src/utils/layerBuilders/expandPolygonFeatures.ts
